# swagger.yaml
openapi: 3.0.0
info:
  title: LCMS API
  description: Tài liệu API cho hệ thống Quản lý Trung tâm Học tập (LCMS)
  version: 2.0.0
servers:
  - url: http://localhost:3003/api/v1
    description: Local Development Server
security:
  - bearerAuth: []
...
paths:
...
  /teachers:
    post:
      tags: [Teacher]
      summary: Tạo giáo viên mới — BO/STAFF (T.1)
      operationId: createTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherRequest'
      responses:
        '201':
          description: Tạo giáo viên thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Teacher'
        '409': { $ref: '#/components/responses/ConflictError' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      tags: [Teacher]
      summary: Danh sách giáo viên — BO/STAFF (T.2)
      operationId: getTeachers
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: search
          in: query
          description: Tìm kiếm theo tên
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: string, enum: ['true', 'false'] }
        - name: subject
          in: query
          description: Lọc theo môn học
          schema: { type: string }
      responses:
        '200':
          description: Danh sách giáo viên
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Teacher'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
  /teachers/{id}:
    get:
      tags: [Teacher]
      summary: Chi tiết giáo viên — BO/STAFF (T.3)
      operationId: getTeacherById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Chi tiết giáo viên
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Teacher'
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      tags: [Teacher]
      summary: Cập nhật giáo viên — BO/STAFF (T.4)
      operationId: updateTeacher
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeacherRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Teacher'
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      tags: [Teacher]
      summary: Xóa mềm giáo viên — BO/STAFF (T.5)
      operationId: deleteTeacher
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Xóa thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted: { type: boolean }
                          teacherId: { type: string }
        '400': { $ref: '#/components/responses/BadRequestError' }
        '404': { $ref: '#/components/responses/NotFoundError' }

  # ── Parent (Module P) ────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: "Thành công" }
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
    ApiErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "Lỗi" }
        code: { type: string, example: "NOT_FOUND" }
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        details: { type: object }
    PaginationMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }

    # ── Auth ──────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        _id: { type: string }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string, nullable: true }
        role:
          type: string
          enum: [SYSTEM_OWNER, BRANCH_OWNER, STAFF, TEACHER, STUDENT, PARENT]
        branchId: { type: string, nullable: true }
        userCode: { type: string }
        avatarUrl: { type: string, nullable: true }
        dateOfBirth: { type: string, format: date-time, nullable: true }
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TeacherInfo:
      type: object
      properties:
        subjects:
          type: array
          items: { type: string }
        joinDate: { type: string, format: date, nullable: true }
        activeClassIds:
          type: array
          items: { type: string }
    Teacher:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            teacherInfo:
              $ref: '#/components/schemas/TeacherInfo'
...
    CreateTeacherRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName: { type: string, minLength: 1 }
        email: { type: string, format: email }
        phone: { type: string }
        password:
          type: string
          minLength: 8
          description: Phải có chữ hoa, chữ thường, số và ký tự đặc biệt
        dateOfBirth: { type: string, format: date }
        gender:
          type: string
          enum: [male, female, other]
        teacherInfo:
          type: object
          properties:
            subjects:
              type: array
              items: { type: string }
            joinDate: { type: string, format: date }
    UpdateTeacherRequest:
      type: object
      properties:
        fullName: { type: string, minLength: 1 }
        phone: { type: string }
        dateOfBirth: { type: string, format: date }
        gender:
          type: string
          enum: [male, female, other]
        avatarUrl: { type: string, format: uri }
        teacherInfo:
          type: object
          properties:
            subjects:
              type: array
              items: { type: string }
            joinDate: { type: string, format: date }

    # ── Parent ───────────────────────────────────────────────────
...
  responses:
...
    NotFoundError:
      description: Không tìm thấy tài nguyên
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ConflictError:
      description: Dữ liệu đã tồn tại hoặc xung đột trạng thái
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ValidationError:
      description: Dữ liệu gửi lên không đúng định dạng
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'